MDEV-21856 XID_t::formatID has to be constrained to 4 byte size
Engine (Innodb) and XA replication MDEV-742 requires the XID member
be of a constant minimum across supported platform ulong size which
is 4 bytes.

That is implemented.
Specifically -1 of uint32 is INVALID format id value to mark
xids internally, and xid value that consist of it won't be accepted.
Additionally two more values are reserved for future to make up
the max XID::formatID = (1<<32) - 4.