MDEV-20590 Introduce a file format constraint to ALTER TABLE
If a table is altered using the MDEV-11369/MDEV-15562/MDEV-13134
ALGORITHM=INSTANT, it can force the table to use a non-canonical
format:

* A hidden metadata record at the start of the clustered index
is used to store each column's DEFAULT value. This makes it possible
to add new columns that have default values without rebuilding the table.

* Starting with MDEV-15562 in MariaDB Server 10.4, a BLOB in the
hidden metadata record is used to store column mappings. This makes
it possible to drop or reorder columns without rebuilding the table.
This also makes it possible to add columns to any position or drop
columns from any position in the table without rebuilding the table.

If a column is dropped without rebuilding the table, old records
will contain garbage in that column's former position, and new records
will be written with NULL values, empty strings, or dummy values.

This is generally not a problem. However, there may be cases where
users may want to avoid putting a table into this format.
For example, users may want to ensure that future UPDATE operations
after an ADD COLUMN will be performed in-place, to reduce write
amplification. (Instantly added columns are essentially always
variable-length.) Users might also want to avoid bugs similar to
MDEV-19916, or they may want to be able to export tables to
older versions of the server.

We will revive the previously deprecated option innodb_file_format,
with the following values:

* barracuda (0): do not allow instant add/drop/reorder,
to maintain format compatibility with MariaDB 10.x and MySQL 5.x

* strict_barracuda: Like 'barracuda', but on ALTER TABLE, if a table
(or partition) is not in the canonical format, force a table rebuild.

* append (default in 10.3): Store a hidden metadata record that
allows columns to be appended to the table (MDEV-11369);
compatible with MariaDB 10.3

* strict_append: Like 'append', but on ALTER TABLE, if a table
(or partition) is not in the 10.3-compatible format,
force the table to be rebuilt. (On 10.3, this is the same as 'append'.)

Starting with 10.4:

* mapped (default): Like 'append', but allow the metadata record to
store a column map, to support instant add/drop/reorder (MDEV-15562)
* strict_mapped: Like 'mapped', but on ALTER TABLE, if a table
(or partition) is not in the 10.4-compatible format,
force the table to be rebuilt. (On 10.4, this is the same as 'mapped'.)